fix(workbench): viewing trasnform component shouldn't crash
some latent cleanup here. I ended up doing a bunch of refactoring at the same time. Damn that DatasetComponent needs to drop in size.
