+++ app/components/Transform.tsx
@@ -1,25 +1,35 @@
 import * as React from 'react'
+import { connect } from 'react-redux'
+import { bindActionCreators, Dispatch } from 'redux'
 
-import { WorkingDataset } from '../models/store'
-import { ApiActionThunk } from '../store/api'
-
+import { Transform } from '../models/dataset'
+import Store from '../models/store'
+import { QriRef } from '../models/qriRef'
+import { fsiWrite } from '../actions/api'
 import Code from './Code'
 
 export interface TransformProps {
-  peername: string
-  name: string
-  value: string
-  preview: string
-  history: boolean
-  workingDataset: WorkingDataset
-  write: (peername: string, name: string, dataset: any) => ApiActionThunk
+  data: Transform
+  qriRef: QriRef
+
+  // TODO (b5) - work in progress
+  dryRun?: () => void
 }
 
-// TODO (b5) - setting a default value here b/c we're getting null value from
-// somewhere when we click workbench/transform component. need to investigate
-// and not pass a nonexistent value, ever.
-const Transform: React.FunctionComponent<TransformProps> = ({ value = '' }) => {
-  return <Code data={value} />
+export const TransformComponent: React.FunctionComponent<TransformProps> = ({ data }) => {
+  return <Code data={data.scriptBytes || ''} />
 }
 
-export default Transform
+const mapStateToProps = (state: Store, ownProps: TransformProps) => {
+  return ownProps
+}
+
+const mergeProps = (props: any, actions: any): TransformProps => { //eslint-disable-line
+  return { ...props, ...actions }
+}
+
+const mapDispatchToProps = (dispatch: Dispatch) => {
+  return bindActionCreators({ write: fsiWrite }, dispatch)
+}
+
+export default connect(mapStateToProps, mapDispatchToProps, mergeProps)(TransformComponent)
